rename packages, remove module common-config
