--- fahrschein-starter-project/fahrschein-spring-boot-starter-config/src/main/java/org/zalando/spring/boot/nakadi/config/JitterType.java
@@ -1,5 +1,0 @@
-package org.zalando.spring.boot.nakadi.config;
-
-public enum JitterType {
-    EQUAL,FULL;
-}
