+++ fahrschein-starter-project/fahrschein-spring-boot-starter-config/src/main/java/org/zalando/spring/boot/nakadi/config/properties/HttpConfig.java
@@ -1,4 +1,4 @@
-package org.zalando.spring.boot.nakadi.config;
+package org.zalando.spring.boot.nakadi.config.properties;
 
 import java.util.concurrent.TimeUnit;
