+++ fahrschein-starter-project/fahrschein-spring-boot-starter-config/src/main/java/org/zalando/spring/boot/nakadi/config/properties/JitterType.java
@@ -1,0 +1,5 @@
+package org.zalando.spring.boot.nakadi.config.properties;
+
+public enum JitterType {
+    EQUAL,FULL;
+}
