+++ fahrschein-starter-project/fahrschein-spring-boot-starter-spring-boot-1-support/src/main/java/org/zalando/spring/boot/nakadi/config/SpringBoot1SettingsParser.java
@@ -3,6 +3,8 @@ package org.zalando.spring.boot.nakadi.config;
 import org.springframework.boot.bind.PropertiesConfigurationFactory;
 import org.springframework.core.env.ConfigurableEnvironment;
 import org.springframework.util.ClassUtils;
+import org.zalando.spring.boot.nakadi.config.properties.FahrscheinConfigProperties;
+import org.zalando.spring.boot.nakadi.config.properties.SettingsParser;
 
 import lombok.SneakyThrows;
