+++ fahrschein-starter-project/fahrschein-spring-boot-starter-spring-boot-1-support/src/main/resources/META-INF/services/org.zalando.spring.boot.nakadi.config.properties.SettingsParser
