+++ fahrschein-starter-project/fahrschein-spring-boot-starter-spring-boot-2-support/pom.xml
@@ -22,7 +22,6 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot</artifactId>
-            <version>2.1.7.RELEASE</version>
             <scope>compile</scope>
             <optional>true</optional>
         </dependency>
