+++ fahrschein-starter-project/fahrschein-spring-boot-starter/src/main/java/org/zalando/spring/boot/nakadi/config/CursorManagerFactory.java
@@ -5,6 +5,7 @@ import java.net.URI;
 import org.zalando.fahrschein.CursorManager;
 import org.zalando.fahrschein.ManagedCursorManager;
 import org.zalando.fahrschein.http.api.RequestFactory;
+import org.zalando.spring.boot.nakadi.config.properties.AbstractConfig;
 
 class CursorManagerFactory {
