+++ fahrschein-starter-project/fahrschein-spring-boot-starter/src/main/java/org/zalando/spring/boot/nakadi/config/NakadiConsumer.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 
 import org.zalando.fahrschein.IORunnable;
 import org.zalando.spring.boot.nakadi.NakadiListener;
+import org.zalando.spring.boot.nakadi.config.properties.ConsumerConfig;
 
 interface NakadiConsumer {
